+++ Traincrew_MultiATS_Server/Hubs/TrainHub.cs
@@ -18,7 +18,8 @@ public class TrainHub(
     SignalService signalService,
     OperationNotificationService operationNotificationService,
     ProtectionService protectionService,
-    RendoService rendoService) : Hub<ITrainClientContract>, ITrainHubContract
+    RendoService rendoService,
+    RouteService routeService) : Hub<ITrainClientContract>, ITrainHubContract
 {
     public async Task<DataFromServer> SendData_ATS(DataToServer clientData)
     {
@@ -78,7 +79,7 @@ public class TrainHub(
             Name = pair.Key,
             phase = pair.Value
         }).ToList();
-        serverData.RouteData = await rendoService.GetActiveRoutes();
+        serverData.RouteData = await routeService.GetActiveRoutes();
         return serverData;
     }
 }
