+++ Traincrew_MultiATS_Server/Services/RendoService.cs
@@ -18,7 +18,6 @@ using Traincrew_MultiATS_Server.Repositories.SwitchingMachineRoute;
 using Traincrew_MultiATS_Server.Repositories.ThrowOutControl;
 using Traincrew_MultiATS_Server.Repositories.TrackCircuit;
 using Route = Traincrew_MultiATS_Server.Models.Route;
-using RouteData = Traincrew_MultiATS_Server.Common.Models.RouteData;
 
 namespace Traincrew_MultiATS_Server.Services;
 
@@ -1463,25 +1462,5 @@ public class RendoService(
         List<LockCondition> lockConditions)
     {
         return lockConditions.OfType<LockConditionObject>().Select(lc => lc.ObjectId).ToList();
-    }
-
-    private static RouteData ToRouteData(Route route)
-    {
-        // Todo: 後でこいつはRouteSerivceとして分離する
-        return new()
-        {
-            TcName = route.TcName,
-            RouteType = route.RouteType,
-            RootId = route.RootId,
-            Indicator = route.Indicator,
-            ApproachLockTime = route.ApproachLockTime,
-        };
-    }
-
-    public async Task<List<RouteData>> GetActiveRoutes()
-    {
-        var routeIds = await routeRepository.GetIdsWhereLeverRelayIsRaised();
-        var routes = await routeRepository.GetByIdsWithState(routeIds);
-        return routes.Select(ToRouteData).ToList();
     }
 }
